src/requirements.txt
# Google Cloud libraries
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4
google-cloud-bigquery==3.14.1
google-auth==2.26.2
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
pyxlsb==1.0.10  # For reading .xlsb (Excel Binary Workbook) files
xlrd==2.0.1  # For reading .xls (Excel 97-2003) files
numpy==1.26.2
ijson==3.2.3  # Streaming JSON parser for large files

# Date Parsing
python-dateutil==2.8.2
pytz==2023.3.post1
types-pytz==2023.3.1.1
types-python-dateutil==2.8.19.14

# Email Parsing
email-validator==2.1.0.post1

# Utilities
functions-framework==3.5.0
Flask==3.0.0
python-dotenv==0.21.0

# Logging and Monitoring
structlog==23.2.0
google-cloud-logging==3.5.0

# Testing (for development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Browser Automation
playwright==1.40.0
keyring==24.3.0

# HTTP clients
httpx==0.27.2
aiohttp==3.9.1
